DS/requirements.txt
matplotlib
numpy
pillow
PyQt5